scripts/requirements.txt
python-docx>=1.1,<1.3
mcp[cli]>=1.0
